Refactored system tests into a standalone test library and test drivers.
